requirements.txt
streamlit==1.57.0
chromadb==1.5.8
sentence-transformers==5.4.1
pypdf==6.10.2
python-docx==1.2.0
